+++ packages/ap-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @recative/ap-core
 
+## 0.2.3
+
+### Patch Changes
+
+- e7b5ff1: fix: Texture update timing
+
+## 0.2.2
+
+### Patch Changes
+
+- 39dbb03: fix: Update texture scale
+
 ## 0.2.1
 
 ### Patch Changes